+++ test/fixtures/doublePlaces.mbtiles
