+++ test/fixtures/doubledPOI.mbtiles
