+++ validators/doubledPOI/README.md
@@ -1,6 +1,6 @@
 #### Description
 
-This validator detects cities that have the same attributes.
+This validator detects all polygons that have the same name that a near point. Nodes must be deleted.  
 
 ![image](https://cloud.githubusercontent.com/assets/10425629/21997330/c061b13e-dbfc-11e6-9b72-ee94251effce.png)
 
@@ -8,4 +8,4 @@ This validator detects cities that have the same attributes.
 #### Usage
 
 1. Download [OSM QA Tiles](https://osmlab.github.io/osm-qa-tiles/) for the planet or a country that you are interested in. 
-2. For example, to run for peru.mbtiles: `osmlint doublecities --bbox="[-85.348663,-19.664209,-64.606476,1.3831570]" --zoom=12 peru.mbtiles`
+2. For example, to run for peru.mbtiles: `osmlint doublepoi --bbox="[-85.348663,-19.664209,-64.606476,1.3831570]" --zoom=12 peru.mbtiles`
