+++ validators/doubledPOI/index.js
